Misc refactoring, update CI, add example data
