+++ .gitignore
@@ -1,3 +1,2 @@
 /target/
 /.local/
-/VERSION
