+++ example-data/LIST-VAR-1.txt
@@ -1,0 +1,43 @@
+BEGIN LIST VAR alpha
+battery.charge: 97
+battery.charge.low: 20
+battery.charge.warning: 20
+battery.runtime: 1693
+battery.temperature: 29.9
+battery.type: PbAc
+battery.voltage: 40.1
+battery.voltage.nominal: 40.0
+device.mfr: Hewlett-Packard
+device.model: HP T1500 G3 UPS
+device.part: DF321A
+device.serial: HIDDEN
+device.type: ups
+input.frequency: 50.0
+input.transfer.high: 266.0
+input.transfer.low: 194.0
+input.voltage: 230.0
+input.voltage.nominal: 230
+output.current: 1.3
+output.frequency: 50.0
+output.frequency.nominal: 50
+output.voltage: 230.0
+output.voltage.nominal: 230
+ups.beeper.status: enabled
+ups.delay.shutdown: 20
+ups.delay.start: 30
+ups.firmware: 100
+ups.load: 31
+ups.load.nominal: 80
+ups.mfr: Hewlett-Packard
+ups.model: HP T1500 G3 UPS
+ups.power: 302.0
+ups.power.nominal: 1400
+ups.productid: 4de5
+ups.serial: HIDDEN
+ups.status: OL CHRG
+ups.test.result: Done and passed
+ups.timer.reboot: -1
+ups.timer.shutdown: -1
+ups.timer.start: -1
+ups.vendorid: 03f0
+END LIST VAR alpha
