+++ example-data/LIST-VAR-2.txt
@@ -1,0 +1,54 @@
+BEGIN LIST VAR ups-2
+VAR apcups01 ambient.1.humidity.alarm.high "60.00"
+VAR apcups01 ambient.1.humidity.alarm.low "30.00"
+VAR apcups01 ambient.1.temperature.alarm.high "40.00"
+VAR apcups01 ambient.1.temperature.alarm.low "10.00"
+VAR apcups01 battery.charge "100.00"
+VAR apcups01 battery.charge.restart "15"
+VAR apcups01 battery.current "0.00"
+VAR apcups01 battery.date "10/06/15"
+VAR apcups01 battery.packs "0.00"
+VAR apcups01 battery.packs.bad "0.00"
+VAR apcups01 battery.runtime "9180.00"
+VAR apcups01 battery.runtime.low "120"
+VAR apcups01 battery.voltage "27.50"
+VAR apcups01 battery.voltage.nominal "0.00"
+VAR apcups01 device.mfr "APC"
+VAR apcups01 device.model "Smart-UPS 1000"
+VAR apcups01 device.serial "MASKED"
+VAR apcups01 device.type "ups"
+VAR apcups01 driver.name "snmp-ups"
+VAR apcups01 driver.parameter.pollfreq "10"
+VAR apcups01 driver.parameter.pollinterval "2"
+VAR apcups01 driver.parameter.port "192.168.1.9"
+VAR apcups01 driver.parameter.snmp_version "v2c"
+VAR apcups01 driver.parameter.synchronous "no"
+VAR apcups01 driver.version "2.7.4"
+VAR apcups01 driver.version.data "apcc MIB 1.2"
+VAR apcups01 driver.version.internal "0.97"
+VAR apcups01 input.frequency "50.00"
+VAR apcups01 input.sensitivity "medium"
+VAR apcups01 input.transfer.high "253"
+VAR apcups01 input.transfer.low "208"
+VAR apcups01 input.transfer.reason "selfTest"
+VAR apcups01 input.voltage "234.70"
+VAR apcups01 input.voltage.maximum "236.10"
+VAR apcups01 input.voltage.minimum "233.20"
+VAR apcups01 output.current "0.00"
+VAR apcups01 output.frequency "50.00"
+VAR apcups01 output.voltage "234.70"
+VAR apcups01 output.voltage.nominal "220"
+VAR apcups01 ups.delay.shutdown "90"
+VAR apcups01 ups.delay.start "0"
+VAR apcups01 ups.firmware "600.3.I"
+VAR apcups01 ups.id "UPS_1K"
+VAR apcups01 ups.load "5.80"
+VAR apcups01 ups.mfr "APC"
+VAR apcups01 ups.mfr.date "04/23/05"
+VAR apcups01 ups.model "Smart-UPS 1000"
+VAR apcups01 ups.serial "MASKED"
+VAR apcups01 ups.status "OL"
+VAR apcups01 ups.temperature "28.80"
+VAR apcups01 ups.test.date "03/11/2021"
+VAR apcups01 ups.test.result "Ok"
+END LIST VAR ups-2
