+++ example-data/LIST-VAR-3.txt
@@ -1,0 +1,43 @@
+BEGIN LIST VAR ups-3
+VAR bagien battery.alarm.threshold "0"
+VAR bagien battery.charge "100.0"
+VAR bagien battery.charge.restart "00"
+VAR bagien battery.date "01/09/15"
+VAR bagien battery.runtime "3300"
+VAR bagien battery.runtime.low "120"
+VAR bagien battery.voltage "13.85"
+VAR bagien battery.voltage.nominal "012"
+VAR bagien device.mfr "APC"
+VAR bagien device.model "Smart-UPS 620   "
+VAR bagien device.serial "MASKED"
+VAR bagien device.type "ups"
+VAR bagien driver.name "apcsmart"
+VAR bagien driver.parameter.pollinterval "2"
+VAR bagien driver.parameter.port "/dev/ttyS0"
+VAR bagien driver.parameter.synchronous "no"
+VAR bagien driver.version "2.7.4"
+VAR bagien driver.version.internal "3.1"
+VAR bagien input.frequency "50.00"
+VAR bagien input.quality "FF"
+VAR bagien input.sensitivity "H"
+VAR bagien input.transfer.high "253"
+VAR bagien input.transfer.low "208"
+VAR bagien input.transfer.reason "unacceptable utility voltage rate of change"
+VAR bagien input.voltage "230.4"
+VAR bagien input.voltage.maximum "230.4"
+VAR bagien input.voltage.minimum "228.9"
+VAR bagien output.voltage "230.4"
+VAR bagien output.voltage.nominal "230"
+VAR bagien ups.delay.shutdown "020"
+VAR bagien ups.delay.start "000"
+VAR bagien ups.firmware "22.6.I"
+VAR bagien ups.id "BAGIEN  "
+VAR bagien ups.load "014.9"
+VAR bagien ups.mfr "APC"
+VAR bagien ups.mfr.date "09/20/01"
+VAR bagien ups.model "Smart-UPS 620   "
+VAR bagien ups.serial "MASKED"
+VAR bagien ups.status "OL"
+VAR bagien ups.test.interval "1209600"
+VAR bagien ups.test.result "NO"
+END LIST VAR ups-3
