+++ example-data/LIST-VAR-4.txt
@@ -1,0 +1,43 @@
+BEGIN LIST VAR ups-4
+VAR alpha battery.charge "100"
+VAR alpha battery.charge.low "10"
+VAR alpha battery.charge.warning "20"
+VAR alpha battery.mfr.date "1"
+VAR alpha battery.runtime "1320"
+VAR alpha battery.runtime.low "300"
+VAR alpha battery.type "PbAcid"
+VAR alpha battery.voltage "260.0"
+VAR alpha battery.voltage.nominal "120"
+VAR alpha device.mfr "1"
+VAR alpha device.model "2200R"
+VAR alpha device.serial "HIDDEN"
+VAR alpha device.type "ups"
+VAR alpha driver.name "usbhid-ups"
+VAR alpha driver.parameter.offdelay "60"
+VAR alpha driver.parameter.ondelay "120"
+VAR alpha driver.parameter.pollfreq "30"
+VAR alpha driver.parameter.pollinterval "2"
+VAR alpha driver.parameter.port "auto"
+VAR alpha driver.parameter.synchronous "no"
+VAR alpha driver.version "2.7.4"
+VAR alpha driver.version.data "CyberPower HID 0.4"
+VAR alpha driver.version.internal "0.41"
+VAR alpha input.transfer.high "290"
+VAR alpha input.transfer.low "165"
+VAR alpha input.voltage "238.7"
+VAR alpha input.voltage.nominal "230"
+VAR alpha output.voltage "237.2"
+VAR alpha ups.beeper.status "enabled"
+VAR alpha ups.delay.shutdown "60"
+VAR alpha ups.delay.start "120"
+VAR alpha ups.load "21"
+VAR alpha ups.mfr "1"
+VAR alpha ups.model "2200R"
+VAR alpha ups.productid "0601"
+VAR alpha ups.realpower.nominal "2200"
+VAR alpha ups.serial "HIDDEN"
+VAR alpha ups.status "OL"
+VAR alpha ups.timer.shutdown "-60"
+VAR alpha ups.timer.start "-60"
+VAR alpha ups.vendorid "0764"
+END LIST VAR ups-4
