--- manage/build-run.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-set -eu
-
-manage/build.sh
-
-manage/run.sh
