+++ manage/build.sh
@@ -4,7 +4,4 @@ set -eu
 
 DC="docker-compose -f manage/docker-compose.yml"
 
-# Add placeholder version file
-echo "0.0.0-SNAPSHOT" > VERSION
-
 $DC build
