+++ manage/docker-compose.yml
@@ -14,8 +14,9 @@ services:
       - HTTP_PORT=9995
       - HTTP_PATH=/nut
       - LOG_REQUESTS_CONSOLE=true
-    ports:
-      - "127.0.0.1:9995:9995/tcp"
+    # ports:
+    #   - "127.0.0.1:9995:9995/tcp"
+    network_mode: host
 
   # prometheus:
   #   image: prom/prometheus:latest
