+++ src/meta.rs
@@ -1,0 +1,4 @@
+pub const APP_NAME: &str = "Prometheus NUT Exporter";
+pub const APP_AUTHOR: &str = "HON95";
+// Automatically set during build
+pub const APP_VERSION: &str = "0.0.0-SNAPSHOT";
